+++ activerecord/test/cases/adapters/abstract_mysql_adapter/mysql_explain_test.rb
@@ -22,17 +22,17 @@ class MySQLExplainTest < ActiveRecord::AbstractMysqlTestCase
   end
 
   def test_explain_with_options_as_symbol
-    explain = Author.where(id: 1).explain(explain_option)
+    explain = Author.where(id: 1).explain(options: [explain_option])
     assert_match %(#{expected_analyze_clause} SELECT `authors`.* FROM `authors` WHERE `authors`.`id` = 1), explain
   end
 
   def test_explain_with_options_as_strings
-    explain = Author.where(id: 1).explain(explain_option.to_s.upcase)
+    explain = Author.where(id: 1).explain(options: [explain_option.to_s.upcase])
     assert_match %(#{expected_analyze_clause} SELECT `authors`.* FROM `authors` WHERE `authors`.`id` = 1), explain
   end
 
   def test_explain_options_with_eager_loading
-    explain = Author.where(id: 1).includes(:posts).explain(explain_option)
+    explain = Author.where(id: 1).includes(:posts).explain(options: [explain_option])
     assert_match %(#{expected_analyze_clause} SELECT `authors`.* FROM `authors` WHERE `authors`.`id` = 1), explain
     assert_match %(#{expected_analyze_clause} SELECT `posts`.* FROM `posts` WHERE `posts`.`author_id` = 1), explain
   end